+++ spec/models/stellar_base/account_subscription_spec.rb
@@ -1,0 +1,7 @@
+require 'rails_helper'
+
+module StellarBase
+  RSpec.describe AccountSubscription, type: :model do
+    pending "add some examples to (or delete) #{__FILE__}"
+  end
+end