+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Add `StellarBase.on_account_event` and `.subscribe_to_accounts`
+
 ## [0.7.1]
 ### Added
 - Add all attributes for a complete stellar_toml file